UPSTREAM: wacom_usb: Read current firmware index before finding image to write
The value of `self->firmware_index` is not valid until we call
`fu_wac_device_ensure_firmware_index`. Because of this, we would always
request the firmware for index 0. If that slot is currently active then
our attempt to program it will fail since the pages are write-protected.
Annoyingly the only signs that something has gone wrong are that the
flash completes almost instantly and the firmware version never changes.

To fix this we re-order our usage of the variable to be after it
becomes valid. We also add a noisy failure message that is triggered
if no blocks are written.

Fixes: 7afd7cba0d32 ("Use FuFirmware as a container for firmware images")

(cherry picked from commit 1ddac4447b582aed57fd445d76cfb8c064c10641)

BUG=b:189184819
TEST=CQ

Change-Id: I432207b1820c89acca543742b446464799962d0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/fwupd/+/2910044
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Daniel Campello <campello@chromium.org>
Tested-by: Daniel Campello <campello@chromium.org>
Commit-Queue: Daniel Campello <campello@chromium.org>